// src/app/auth/interfaces/post.ts
// Generated by https://quicktype.io
export interface PostRequest {
    title: string;
    description: string;
    content: string;
    tags: string;
    categoria_id: number;
    published: number;
    imagen: string;
    anuncios: string;
}
// Generated by https://quicktype.io

export interface Post {
    id: number;
    user: User;
    categoria: Categoria;
    title: string;
    slug: string;
    description: string;
    content: string;
    imagen: string;
    published: boolean;
    tags: Tag[];
    anuncios: Anuncio[];
    created_at: string;
}

export interface Anuncio {
    id: number;
    name: string;
    image: string;
    categoria_anuncio: Categoria;
    description: string;
    address: string;
    phone: string;
    cellphone: string;
    status: number;
    tags: Tag[];
    created_at: string;
}

export interface Categoria {
    id: number;
    name: string;
    color: string;
    description: string;
    created_at: string;
    image?: string;
}

export interface Tag {
    id: number;
    name: string;
}

export interface User {
    id: number;
    name: string;
    email: string;
    rol: string;
    created_at: string;
}